+++ template/bison/yacc.c
@@ -582,6 +582,39 @@ static const <%= output.int_type_for(output.context.yyr2) %> yyr2[] =
 <%= output.int_array_to_string(output.context.yyr2) %>
 };
 
+<%- if output.pslr_enabled? -%>
+<%= output.pslr_function_declarations %>
+<%- end -%>
+
+<%- if output.pslr_enabled? -%>
+<%= output.pslr_tables_and_functions %>
+<%- end -%>
+
+<%- if output.pslr_enabled? -%>
+#ifndef YYSETSTATE_CONTEXT
+# define YYSETSTATE_CONTEXT(CurrentState) ((void) 0)
+#endif
+<%- end -%>
+
+int
+yy_state_accepts_token (int yystate, int yychar)
+{
+  yysymbol_kind_t yytoken = YYTRANSLATE (yychar);
+  int yyn = yypact[yystate];
+
+  if (yypact_value_is_default (yyn))
+    return 0;
+
+  yyn += yytoken;
+  if (yyn < 0 || YYLAST < yyn || yycheck[yyn] != yytoken)
+    return 0;
+
+  yyn = yytable[yyn];
+  if (yyn <= 0)
+    return !yytable_value_is_error (yyn);
+
+  return 1;
+}
 
 enum { YYENOMEM = -2 };
 
@@ -1582,6 +1615,9 @@ yysetstate:
   YY_IGNORE_USELESS_CAST_BEGIN
   *yyssp = YY_CAST (yy_state_t, yystate);
   YY_IGNORE_USELESS_CAST_END
+<%- if output.pslr_enabled? -%>
+  YYSETSTATE_CONTEXT (yystate);
+<%- end -%>
   YY_STACK_PRINT (yyss, yyssp<%= output.user_args %>);
 
   if (yyss + yystacksize - 1 <= yyssp)
@@ -2065,4 +2101,3 @@ yyreturnlab:
 #line <%= output.aux.epilogue_first_lineno - 1 %> "<%= output.grammar_file_path %>"
 <%= output.aux.epilogue -%>
 <%- end -%>
-